Moved component directory holding SystemStatusModal to client/modules/workspace/src/ and updated corresponding imports and exports
